Fix IntegrationFlow <-> IntFlowDefinition tangles
